[NES] [Breaking] Make NES targets PRG-ROM LMAs accurate.
Previously, the NES targets used a fictious notion of LMA correpsonding
to offsets within the NES ROM image. The intended use of linker LMAs is
to be the visible address of the ROM at start; the corresponding VMA
is intended to be the address of the value at run time. If they differ,
a copy must occur at start.

Using fictitious LMAs meant we were fighting the linker; we needed
complicated expressions to generate the fake LMAs, and duplicated
sections to make sure the fake and real LMAs ligned up.

Instead, we can just let LMAs be LMAs and VMAs be VMAs. We don't
actually *need* an offset in the ROM image; the existing output format
scripts can be generalized to accomodate the existing mappers. This
generatlly reduces the complexity of the NES targets.

As a result, there's a slight breaking change; the MMC1 mapper now uses
prg-rom-fixed.ld in cases where the mapper's address line is
disconnected, and prg-rom-32.ld now refers to the case where it is
connected, but only 32KiB of PRG-ROM is available.
