+++ mos-platform/nes-mmc1/CMakeLists.txt
@@ -6,12 +6,9 @@ endif()
 
 install(FILES
   _prg-rom-banked.ld
-  _prg-rom-fixed.ld
   common.ld
   c-in-prg-ram-0.ld
-  prg-rom-16.ld
   prg-rom-32.ld
-  prg-rom-32-banked.ld
   prg-rom-64.ld
   prg-rom-128.ld
   prg-rom-256.ld
@@ -20,9 +17,9 @@ install(FILES bank.h TYPE INCLUDE)
 
 add_platform_object_file(nes-mmc1-crt0-o crt0.o ines.s)
 add_platform_object_file(nes-mmc1-reset-o reset.o reset.s)
-add_platform_object_file(nes-mmc1-init-prg-rom-32-banked-o
-  init-prg-rom-32-banked.o
-  init-prg-rom-32-banked.s)
+add_platform_object_file(nes-mmc1-init-prg-rom-32-o
+  init-prg-rom-32.o
+  init-prg-rom-32.s)
 add_platform_object_file(nes-mmc1-init-prg-ram-0-o
   init-prg-ram-0.o init-prg-ram-0.s)
