+++ mos-platform/nes-mmc1/c-in-prg-ram-0.ld
@@ -1,5 +1,5 @@
 /* Place C sections in PRG-RAM bank 0. */
-REGION_ALIAS("c_ram_vma", prg_ram_0_vma)
+REGION_ALIAS("c_ram", prg_ram_0)
 
 /* Ensure that PRG-RAM bank 0 is selected before C starts. */
 INPUT(init-prg-ram-0.o)
