--- mos-platform/nes-mmc1/prg-rom-16.ld
@@ -1,8 +1,0 @@
-__prg_rom_size = 16;
-
-MEMORY {
-  /* PRG-ROM VMA. */
-  prg_rom_vma : ORIGIN = 0xc000, LENGTH = 0x4000
-}
-
-INCLUDE _prg-rom-fixed.ld
