+++ mos-platform/nes-mmc1/prg-rom-64.ld
@@ -2,8 +2,11 @@ INCLUDE _prg-rom-banked.ld
 
 __prg_rom_size = 64;
 
-SECTIONS {
-  .reset_0 : OVERLAY { KEEP(*(.reset_0)) } >reset_0
-  .reset_1 : OVERLAY { KEEP(*(.reset_1)) } >reset_1
-  .reset_2 : OVERLAY { KEEP(*(.reset_2)) } >reset_2
+OUTPUT_FORMAT {
+  INCLUDE ines-header.ld
+  FULL(prg_rom_0)     FULL(reset)
+  FULL(prg_rom_1)     FULL(reset)
+  FULL(prg_rom_2)     FULL(reset)
+  FULL(prg_rom_fixed) FULL(reset_fixed)
+  FULL(chr_rom)
 }
