+++ mos-platform/nes-mmc3/prg-rom-banked-8.ld
@@ -1,85 +1,144 @@
-ASSERT(__prg_rom_size > 32,
-       "For less than 32KiB of PRG-ROM, use fixed banking.");
+PROVIDE(__prg_rom_0  = 0x8000);
+PROVIDE(__prg_rom_1  = 0x8000);
+PROVIDE(__prg_rom_2  = 0x8000);
+PROVIDE(__prg_rom_3  = 0x8000);
+PROVIDE(__prg_rom_4  = 0x8000);
+PROVIDE(__prg_rom_5  = 0x8000);
+PROVIDE(__prg_rom_6  = 0x8000);
+PROVIDE(__prg_rom_7  = 0x8000);
+PROVIDE(__prg_rom_8  = 0x8000);
+PROVIDE(__prg_rom_9  = 0x8000);
+PROVIDE(__prg_rom_10 = 0x8000);
+PROVIDE(__prg_rom_11 = 0x8000);
+PROVIDE(__prg_rom_12 = 0x8000);
+PROVIDE(__prg_rom_13 = 0x8000);
+PROVIDE(__prg_rom_14 = 0x8000);
+PROVIDE(__prg_rom_15 = 0x8000);
+PROVIDE(__prg_rom_16 = 0x8000);
+PROVIDE(__prg_rom_17 = 0x8000);
+PROVIDE(__prg_rom_18 = 0x8000);
+PROVIDE(__prg_rom_19 = 0x8000);
+PROVIDE(__prg_rom_20 = 0x8000);
+PROVIDE(__prg_rom_21 = 0x8000);
+PROVIDE(__prg_rom_22 = 0x8000);
+PROVIDE(__prg_rom_23 = 0x8000);
+PROVIDE(__prg_rom_24 = 0x8000);
+PROVIDE(__prg_rom_25 = 0x8000);
+PROVIDE(__prg_rom_26 = 0x8000);
+PROVIDE(__prg_rom_27 = 0x8000);
+PROVIDE(__prg_rom_28 = 0x8000);
+PROVIDE(__prg_rom_29 = 0x8000);
+PROVIDE(__prg_rom_30 = 0x8000);
+PROVIDE(__prg_rom_31 = 0x8000);
+PROVIDE(__prg_rom_32 = 0x8000);
+PROVIDE(__prg_rom_33 = 0x8000);
+PROVIDE(__prg_rom_34 = 0x8000);
+PROVIDE(__prg_rom_35 = 0x8000);
+PROVIDE(__prg_rom_36 = 0x8000);
+PROVIDE(__prg_rom_37 = 0x8000);
+PROVIDE(__prg_rom_38 = 0x8000);
+PROVIDE(__prg_rom_39 = 0x8000);
+PROVIDE(__prg_rom_40 = 0x8000);
+PROVIDE(__prg_rom_41 = 0x8000);
+PROVIDE(__prg_rom_42 = 0x8000);
+PROVIDE(__prg_rom_43 = 0x8000);
+PROVIDE(__prg_rom_44 = 0x8000);
+PROVIDE(__prg_rom_45 = 0x8000);
+PROVIDE(__prg_rom_46 = 0x8000);
+PROVIDE(__prg_rom_47 = 0x8000);
+PROVIDE(__prg_rom_48 = 0x8000);
+PROVIDE(__prg_rom_49 = 0x8000);
+PROVIDE(__prg_rom_50 = 0x8000);
+PROVIDE(__prg_rom_51 = 0x8000);
+PROVIDE(__prg_rom_52 = 0x8000);
+PROVIDE(__prg_rom_53 = 0x8000);
+PROVIDE(__prg_rom_54 = 0x8000);
+PROVIDE(__prg_rom_55 = 0x8000);
+PROVIDE(__prg_rom_56 = 0x8000);
+PROVIDE(__prg_rom_57 = 0x8000);
+PROVIDE(__prg_rom_58 = 0x8000);
+PROVIDE(__prg_rom_59 = 0x8000);
+PROVIDE(__prg_rom_60 = 0x8000);
 
-PROVIDE(__prg_rom_size = 512);
+INCLUDE _prg-rom-banked.ld
 
 /* This bank must be switched to on reset to establish a 24KiB fixed region. */
 __fixed_bank = (__prg_rom_size / 8) - 3;
 
 INPUT(reset-banked-8.o)
 
+__prg_rom_fixed_lma = 0xa000;
+__prg_rom_fixed_offset = __prg_rom_size * 1024 - 0x6000;
 
-MEMORY {
-  prg_rom_fixed_lma : ORIGIN = 0x01000000 + __prg_rom_size * 1024 - 0x6000,
-                      LENGTH = 0x6000
-
-  prg_rom_0_vma     : ORIGIN = 0x008000, LENGTH = 0x2000
-  prg_rom_1_vma     : ORIGIN = 0x018000, LENGTH = 0x2000
-  prg_rom_2_vma     : ORIGIN = 0x028000, LENGTH = 0x2000
-  prg_rom_3_vma     : ORIGIN = 0x038000, LENGTH = 0x2000
-  prg_rom_4_vma     : ORIGIN = 0x048000, LENGTH = 0x2000
-  prg_rom_5_vma     : ORIGIN = 0x058000, LENGTH = 0x2000
-  prg_rom_6_vma     : ORIGIN = 0x068000, LENGTH = 0x2000
-  prg_rom_7_vma     : ORIGIN = 0x078000, LENGTH = 0x2000
-  prg_rom_8_vma     : ORIGIN = 0x088000, LENGTH = 0x2000
-  prg_rom_9_vma     : ORIGIN = 0x098000, LENGTH = 0x2000
-  prg_rom_10_vma    : ORIGIN = 0x0a8000, LENGTH = 0x2000
-  prg_rom_11_vma    : ORIGIN = 0x0b8000, LENGTH = 0x2000
-  prg_rom_12_vma    : ORIGIN = 0x0c8000, LENGTH = 0x2000
-  prg_rom_13_vma    : ORIGIN = 0x0d8000, LENGTH = 0x2000
-  prg_rom_14_vma    : ORIGIN = 0x0e8000, LENGTH = 0x2000
-  prg_rom_15_vma    : ORIGIN = 0x0f8000, LENGTH = 0x2000
-  prg_rom_16_vma    : ORIGIN = 0x108000, LENGTH = 0x2000
-  prg_rom_17_vma    : ORIGIN = 0x118000, LENGTH = 0x2000
-  prg_rom_18_vma    : ORIGIN = 0x128000, LENGTH = 0x2000
-  prg_rom_19_vma    : ORIGIN = 0x138000, LENGTH = 0x2000
-  prg_rom_20_vma    : ORIGIN = 0x148000, LENGTH = 0x2000
-  prg_rom_21_vma    : ORIGIN = 0x158000, LENGTH = 0x2000
-  prg_rom_22_vma    : ORIGIN = 0x168000, LENGTH = 0x2000
-  prg_rom_23_vma    : ORIGIN = 0x178000, LENGTH = 0x2000
-  prg_rom_24_vma    : ORIGIN = 0x188000, LENGTH = 0x2000
-  prg_rom_25_vma    : ORIGIN = 0x198000, LENGTH = 0x2000
-  prg_rom_26_vma    : ORIGIN = 0x1a8000, LENGTH = 0x2000
-  prg_rom_27_vma    : ORIGIN = 0x1b8000, LENGTH = 0x2000
-  prg_rom_28_vma    : ORIGIN = 0x1c8000, LENGTH = 0x2000
-  prg_rom_29_vma    : ORIGIN = 0x1d8000, LENGTH = 0x2000
-  prg_rom_30_vma    : ORIGIN = 0x1e8000, LENGTH = 0x2000
-  prg_rom_31_vma    : ORIGIN = 0x1f8000, LENGTH = 0x2000
-  prg_rom_32_vma    : ORIGIN = 0x208000, LENGTH = 0x2000
-  prg_rom_33_vma    : ORIGIN = 0x218000, LENGTH = 0x2000
-  prg_rom_34_vma    : ORIGIN = 0x228000, LENGTH = 0x2000
-  prg_rom_35_vma    : ORIGIN = 0x238000, LENGTH = 0x2000
-  prg_rom_36_vma    : ORIGIN = 0x248000, LENGTH = 0x2000
-  prg_rom_37_vma    : ORIGIN = 0x258000, LENGTH = 0x2000
-  prg_rom_38_vma    : ORIGIN = 0x268000, LENGTH = 0x2000
-  prg_rom_39_vma    : ORIGIN = 0x278000, LENGTH = 0x2000
-  prg_rom_40_vma    : ORIGIN = 0x288000, LENGTH = 0x2000
-  prg_rom_41_vma    : ORIGIN = 0x298000, LENGTH = 0x2000
-  prg_rom_42_vma    : ORIGIN = 0x2a8000, LENGTH = 0x2000
-  prg_rom_43_vma    : ORIGIN = 0x2b8000, LENGTH = 0x2000
-  prg_rom_44_vma    : ORIGIN = 0x2c8000, LENGTH = 0x2000
-  prg_rom_45_vma    : ORIGIN = 0x2d8000, LENGTH = 0x2000
-  prg_rom_46_vma    : ORIGIN = 0x2e8000, LENGTH = 0x2000
-  prg_rom_47_vma    : ORIGIN = 0x2f8000, LENGTH = 0x2000
-  prg_rom_48_vma    : ORIGIN = 0x308000, LENGTH = 0x2000
-  prg_rom_49_vma    : ORIGIN = 0x318000, LENGTH = 0x2000
-  prg_rom_50_vma    : ORIGIN = 0x328000, LENGTH = 0x2000
-  prg_rom_51_vma    : ORIGIN = 0x338000, LENGTH = 0x2000
-  prg_rom_52_vma    : ORIGIN = 0x348000, LENGTH = 0x2000
-  prg_rom_53_vma    : ORIGIN = 0x358000, LENGTH = 0x2000
-  prg_rom_54_vma    : ORIGIN = 0x368000, LENGTH = 0x2000
-  prg_rom_55_vma    : ORIGIN = 0x378000, LENGTH = 0x2000
-  prg_rom_56_vma    : ORIGIN = 0x388000, LENGTH = 0x2000
-  prg_rom_57_vma    : ORIGIN = 0x398000, LENGTH = 0x2000
-  prg_rom_58_vma    : ORIGIN = 0x3a8000, LENGTH = 0x2000
-  prg_rom_59_vma    : ORIGIN = 0x3b8000, LENGTH = 0x2000
-  prg_rom_60_vma    : ORIGIN = 0x3c8000, LENGTH = 0x2000
-
-  prg_rom_fixed_vma : ORIGIN = 0xa000,   LENGTH = 0x6000
-}
-
-REGION_ALIAS("c_rom_vma", prg_rom_fixed_vma)
-REGION_ALIAS("c_rom_lma", prg_rom_fixed_lma)
+MEMORY { prg_rom_fixed : ORIGIN = __prg_rom_fixed_lma, LENGTH = 0x6000 }
+REGION_ALIAS("c_rom", prg_rom_fixed)
 
 INCLUDE dpcm.ld
 INCLUDE _reset.ld
+
+OUTPUT_FORMAT {
+  INCLUDE ines-header.ld
+  FULL(prg_rom_0)
+  FULL(prg_rom_1)
+  FULL(prg_rom_2)
+  FULL(prg_rom_3)
+  FULL(prg_rom_4)
+  FULL(prg_rom_5)
+  FULL(prg_rom_6)
+  FULL(prg_rom_7)
+  FULL(prg_rom_8)
+  FULL(prg_rom_9)
+  FULL(prg_rom_10)
+  FULL(prg_rom_11)
+  FULL(prg_rom_12)
+  FULL(prg_rom_13)
+  FULL(prg_rom_14)
+  FULL(prg_rom_15)
+  FULL(prg_rom_16)
+  FULL(prg_rom_17)
+  FULL(prg_rom_18)
+  FULL(prg_rom_19)
+  FULL(prg_rom_20)
+  FULL(prg_rom_21)
+  FULL(prg_rom_22)
+  FULL(prg_rom_23)
+  FULL(prg_rom_24)
+  FULL(prg_rom_25)
+  FULL(prg_rom_26)
+  FULL(prg_rom_27)
+  FULL(prg_rom_28)
+  FULL(prg_rom_29)
+  FULL(prg_rom_30)
+  FULL(prg_rom_31)
+  FULL(prg_rom_32)
+  FULL(prg_rom_33)
+  FULL(prg_rom_34)
+  FULL(prg_rom_35)
+  FULL(prg_rom_36)
+  FULL(prg_rom_37)
+  FULL(prg_rom_38)
+  FULL(prg_rom_39)
+  FULL(prg_rom_40)
+  FULL(prg_rom_41)
+  FULL(prg_rom_42)
+  FULL(prg_rom_43)
+  FULL(prg_rom_44)
+  FULL(prg_rom_45)
+  FULL(prg_rom_46)
+  FULL(prg_rom_47)
+  FULL(prg_rom_48)
+  FULL(prg_rom_49)
+  FULL(prg_rom_50)
+  FULL(prg_rom_51)
+  FULL(prg_rom_52)
+  FULL(prg_rom_53)
+  FULL(prg_rom_54)
+  FULL(prg_rom_55)
+  FULL(prg_rom_56)
+  FULL(prg_rom_57)
+  FULL(prg_rom_58)
+  FULL(prg_rom_59)
+  FULL(prg_rom_60)
+  FULL(prg_rom_fixed)
+  FULL(chr_rom)
+}
