+++ mos-platform/nes/c-in-prg-ram-fixed.ld
@@ -1,4 +1,4 @@
 /* Place C sections in PRG-RAM. */
-REGION_ALIAS("c_ram_vma", prg_ram)
+REGION_ALIAS("c_ram", prg_ram)
 
 __stack = ORIGIN(prg_ram) + LENGTH(prg_ram) - 1;
