+++ mos-platform/nes/c-in-ram.ld
@@ -1,3 +1,3 @@
 /* Place C sections in NES RAM. */
-REGION_ALIAS("c_ram_vma", ram)
+REGION_ALIAS("c_ram", ram)
 __stack = 0x07ff;
