+++ mos-platform/nes/dpcm.ld
@@ -4,6 +4,6 @@ SECTIONS {
   .dpcm : ALIGN(64) {
        . = ABSOLUTE(.) < 0xc000 ? 0xc000 : .;
        KEEP(*(.dpcm .dpcm.*))
-  } >c_rom_vma AT>c_rom_lma
+  } >c_rom
   PROVIDE(__dpcm_offset = (ADDR(.dpcm) - 0xc000) >> 6);
 }
