--- mos-platform/nes/ines-roms.ld
@@ -1,2 +1,0 @@
-FULL(prg_rom)
-FULL(chr_rom)
