--- mos-platform/nes/ines.ld
@@ -1,4 +1,0 @@
-OUTPUT_FORMAT {
-  INCLUDE ines-header.ld
-  INCLUDE ines-roms.ld
-}
