+++ mos-platform/nes/prg-rom-fixed.ld
@@ -1,21 +1,23 @@
-/* Common elements of NES-NROM iNES file linker scripts. */
-
 ASSERT(__prg_rom_size == 16 || __prg_rom_size == 32,
-       "NROM only supports 16 KiB or 32 KiB of PRG-ROM")
+       "Fixed PRG-ROM only supports 16 KiB or 32 KiB")
 
-MEMORY {
-  prg_rom_vma (rx) : ORIGIN = __prg_rom_size == 32 ? 0x8000 : 0xC000,
-                     LENGTH = __prg_rom_size * 1024
-}
+__prg_rom_lma    = __prg_rom_size == 32 ? 0x8000 : 0xC000;
+__prg_rom_offset = 0;
 
-REGION_ALIAS("c_rom_vma", prg_rom_vma)
-REGION_ALIAS("c_rom_lma", prg_rom)
+MEMORY { prg_rom (rx) : ORIGIN = __prg_rom_lma, LENGTH = __prg_rom_size * 1024 }
 
-INCLUDE dpcm.ld
+REGION_ALIAS("c_rom", prg_rom)
 
 SECTIONS {
   .prg_rom_fixed : { *(.prg_rom_fixed .prg_rom_fixed.*) } >prg_rom
-  .vector 0xfffa - ORIGIN(prg_rom_vma) + ORIGIN(prg_rom) : {
-    SHORT(nmi) SHORT(_start) SHORT(irq)
-  } >prg_rom
+}
+INCLUDE dpcm.ld
+SECTIONS {
+  .vector 0xfffa : { SHORT(nmi) SHORT(_start) SHORT(irq) } >prg_rom
+}
+
+OUTPUT_FORMAT {
+  INCLUDE ines-header.ld
+  FULL(prg_rom)
+  FULL(chr_rom)
 }
